Move field selector into common folder
