+++ x-pack/plugins/observability/public/pages/slo_edit/components/common/field_selector.stories.tsx
@@ -15,7 +15,7 @@ import { SLO_EDIT_FORM_DEFAULT_VALUES } from '../../constants';
 
 export default {
   component: Component,
-  title: 'app/SLO/EditPage/ApmLatency/FieldSelector',
+  title: 'app/SLO/EditPage/Common/FieldSelector',
   decorators: [KibanaReactStorybookDecorator],
 };
