+++ x-pack/plugins/observability/public/pages/slo_edit/components/common/field_selector.tsx
@@ -52,7 +52,7 @@ export function FieldSelector({
         ? [
             {
               value: '*',
-              label: i18n.translate('xpack.observability.slos.sloEdit.apmFieldSelector.all', {
+              label: i18n.translate('xpack.observability.slos.sloEdit.fieldSelector.all', {
                 defaultMessage: 'All',
               }),
             },
